feat(trace-detail): refactor query service trace apis
